Kernel: Moved all platform dependent code to arch/i686 folder
This is the beginning of several changes that I will make to the kernel, the idea is to separate the code of each architecture as much as possible, so that we can add other platforms (x86_64, aarch64, etc...)

+++ kernel/api/posix/time.h
@@ -2,6 +2,6 @@
 
 struct timespec
 {
-    long long tv_sec;
+    long tv_sec;
     long tv_nsec;
 };
